febio2: changed for-loop index type from size_t to i when loop is inside an omp pragma.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@7457 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioMix/FEBiphasicSoluteDomain.cpp
@@ -244,7 +244,7 @@ void FEBiphasicSoluteDomain::InternalForces(FEGlobalVector& R)
 {
 	size_t NE = m_Elem.size();
 	#pragma omp parallel for shared (NE)
-	for (size_t i=0; i<NE; ++i)
+	for (int i=0; i<NE; ++i)
 	{
 		// element force vector
 		vector<double> fe;
